1-24

+++ src/EnemyMove.java
@@ -2,8 +2,8 @@ public class EnemyMove {
 
     Enemy enemy;
 
-    int xPos;
-    int yPos;
+    double xPos;
+    double yPos;
 
     int routePosX;
     int routePosY;
